requirements_server.txt
flask==3.0.3
apscheduler==3.10.4
requests==2.31.0
pytz==2024.1
google-auth==2.29.0
google-api-python-client==2.128.0
gunicorn==22.0.0
